Create simulation.py
This is my idea for the basis of our program. I haven't finished writing it though. Basically it's a 3d array, where the 2d part is the width and height of the map, and the vertical aspect is the stacking of animals when they move. Since all animals move at once, multiple might end up in one square. I also added some functions to classes like move and reproduce, although I haven't entirely finished them. 

I'm not sure if this is the best way to do it, but it's the way I thought of. And if moving all the animals at once turns out to be too complicated, we can just have one animal at a time move, which would be much simpler.