visualized simulation.py, seems to work but has issues
I began to visualize the code in the main branch. Haven't implemented ui elements from turtletest.py yet but will work on that soon.

Currently my main issues are the following:
First:
it takes forever to draw the grid; over 30 seconds to make a 20x20 and fill it with clonemakers

(perhaps the more important issue):
There seems to be a conflict between the coordinates of the turtles in the array and the turtles being visualized. This might be related to the [0][0] bug, because the turtles being visualized are dispersed just fine across the grid I made rather than being forced into [0][0]. 
Since calculations for game rounds occur based off the array (as they should be), visualized turtles will inexplicably disappear while seemingly being alone because according to their positions on the array they are in the same square as another turtle but on the grid they are elsewhere. I may be biased but due to the bug I am slightly inclined to believe that the positions of the turtles according to the visualization are more accurate than they are according to the array at the moment.