Some Final updates, still requires bugfixes
added some more info and mimic graph
still need to fix turtles moving out the bottom issue

an idea would be to record some generations where there are no bullfrogs simulated, and see if that causes there to be no longer an upward trend in mimicry. This would prove that our simulation results are due to natural predator-prey selection.